backend/requirements.txt
asgiref==3.7.2
black==24.2.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
coreapi==2.3.3
coreschema==0.0.4
coverage==7.4.3
cryptography==42.0.5
dj-database-url==2.1.0
Django==5.0.2
django-cors-headers==4.3.1
django-filter==23.5
django-rest-framework==0.1.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
exceptiongroup==1.2.0
factory-boy==3.3.0
Faker==24.0.0
idna==3.6
inflection==0.5.1
iniconfig==2.0.0
itypes==1.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
mypy-extensions==1.0.0
openapi-codec==1.3.2
packaging==23.2
pathspec==0.12.1
pillow==10.2.0
platformdirs==4.2.0
pluggy==1.4.0
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
pytest==8.0.2
pytest-cov==4.1.0
pytest-django==4.8.0
pytest-factoryboy==2.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2022.7.1
PyYAML==6.0.1
requests==2.31.0
simplejson==3.19.2
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.10.0
uritemplate==4.1.1
urllib3==2.2.1
